ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
pydantic-settings==2.5.2
numpy==1.26.4
scikit-learn==1.5.2
python-multipart==0.0.9
Pillow==10.4.0
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
httpx==0.27.2
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
alembic==1.13.2
python-dotenv==1.0.1
